Fix: 	Stone can do normal move when it must do a eat move. Modified program structure.
