+++ .gitignore
@@ -1,2 +1,3 @@
 *.swp
 .gitingore
+__pycache__
